Enum perfcnt::linux::HardwareEventType

pub enum HardwareEventType {
    CPUCycles,
    Instructions,
    CacheReferences,
    CacheMisses,
    BranchInstructions,
    BranchMisses,
    BusCycles,
    StalledCyclesFrontend,
    StalledCyclesBackend,
    RefCPUCycles,
}

Variants
CPUCycles

Total cycles. Be wary of what happens during CPU frequency scaling.

Instructions

Retired instructions. Be careful, these can be affected by various issues, most notably hardware interrupt counts.

CacheReferences

Cache accesses. Usually this indicates Last Level Cache accesses but this may vary depending on your CPU. This may include prefetches and

CacheMisses

Cache misses. Usually this indicates Last Level Cache misses; this is intended to be used in conjunction with the

BranchInstructions

Retired branch instructions. Prior to Linux 2.6.34, this used the wrong event on AMD processors.

BranchMisses

Mispredicted branch instructions.

BusCycles

Bus cycles, which can be different from total cycles.

StalledCyclesFrontend

Stalled cycles during issue. (Since Linux 3.0)

StalledCyclesBackend

Stalled cycles during retirement. (Since Linux 3.0)

RefCPUCycles

Total cycles; not affected by CPU frequency scaling. (Since Linux 3.3)