Enum perfcnt::linux::ReadFormat

pub enum ReadFormat {
    TotalTimeEnabled,
    TotalTimeRunning,
    FormatId,
    FormatGroup,
}

Variants

TotalTimeEnabled

Adds the 64-bit time_enabled field. This can be used to calculate estimated totals if the PMU is overcommitted and multiplexing is happening.

TotalTimeRunning

Adds the 64-bit time_running field. This can be used to calculate estimated totals if the PMU is overcommitted and multiplexing is happening.

FormatId

Adds a 64-bit unique value that corresponds to the event group.

FormatGroup

Allows all counter values in an event group to be read with one read.