Enum perfcnt::linux::SoftwareEventType

pub enum SoftwareEventType {
    CpuClock,
    TaskClock,
    PageFaults,
    ContextSwitches,
    CpuMigrations,
    PageFaultsMin,
    PageFaultsMaj,
    AlignmentFaults,
    EmulationFaults,
}

Variants
CpuClock

This reports the CPU clock, a high-resolution per-CPU timer.

TaskClock

This reports a clock count specific to the task that is running.

PageFaults

This reports the number of page faults.

ContextSwitches

This counts context switches.

Until Linux 2.6.34, these were all reported as user-space events, after that they are reported as happening in the kernel.

CpuMigrations

This reports the number of times the process has migrated to a new CPU.

PageFaultsMin

This counts the number of minor page faults. These did not require disk I/O to handle.

PageFaultsMaj

This counts the number of major page faults. These required disk I/O to handle.

AlignmentFaults

This counts the number of alignment faults.

These happen when unaligned memory accesses happen; the kernel can handle these but it reduces performance. This happens only on some architectures (never on x86).

(Since Linux 2.6.33)

EmulationFaults

This counts the number of emulation faults. The kernel sometimes traps on unimplemented instructions and emulates them for user space. This can negatively impact performance.

(Since Linux 2.6.33)