Module perfcnt::linux::parser

Uses the nom library to parse the in memory format of perf data structures and transforms them into more rust-like data-strutures.

bin/parse.rs serves as an example program how-to use the parser.

References

The code is inspired by the following articles and existing parser to make sense of the (poorly documented) format: * https://lwn.net/Articles/644919/ * http://man7.org/linux/man-pages/man2/perf_event_open.2.html * https://github.com/andikleen/pmu-tools/tree/master/parser

Current limitations

Only version 2 of the data format

No support for AUX stuff

Sample ID at the end of records is currently ignored

I'm not sure if I'm parsing the BuildId correctly, it seems it can not be recognized

Only support little endian machines

Structs

PerfFile

PerfFileEventDataIter

Functions

parse_comm_record

parse_sample_record