Module perfcnt::linux::parser

Uses the nom library to parse the in memory format of perf data structures and transforms them into more rust-like data-strutures.

References

The code is inspired by the following articles and existing parser to make sense of the (poorly documented) format:

https://lwn.net/Articles/644919/

http://man7.org/linux/man-pages/man2/perf_event_open.2.html

https://github.com/andikleen/pmu-tools/tree/master/parser

Current limitations

Only version 2 of the data format

No support for AUX stuff

Sample ID at the end of records is currently ignored

I'm not sure if I'm parsing the BuildId correctly, it seems it can not be recognized

Only support little endian machines

See also

perf_file.rs -- as an example on how to use the parser function to parse a perf.data file

perf_format.rs -- for all the struct definitions that are parsed here

Functions

parse_branch_entries

parse_branch_entry

parse_build_id_record

parse_comm_record

parse_cpu

parse_cpu_topology

parse_event
Parse an event record.

parse_event_attr

parse_event_desc

parse_event_header

parse_exit_record

parse_file_section

parse_fork_record

parse_group_description

parse_group_descriptions

parse_header

parse_mmap2_record

parse_mmap_record

parse_nrcpus

parse_numa_node

parse_numa_topology

parse_perf_string

parse_perf_string_list

parse_pmu_mapping

parse_pmu_mappings

parse_read_format

parse_read_value

parse_sample_id

parse_sample_record

parse_thread_id

parse_throttle_record

parse_unthrottle_record